+++ enterprise/__init__.py
@@ -2,6 +2,6 @@
 Your project description goes here.
 """
 
-__version__ = "3.27.25"
+__version__ = "3.27.26"
 
 default_app_config = "enterprise.apps.EnterpriseConfig"